modif migration

+++ src/main/resources/db/migrations/V1__init_schema.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS societes_gestion (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nom VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS scpis (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nom VARCHAR(255) NOT NULL,
     minimum_souscription INTEGER,
     capitalisation BIGINT,
@@ -19,14 +19,14 @@ CREATE TABLE IF NOT EXISTS scpis (
     versement_programme BOOLEAN,
     publicite TEXT,
     url_image VARCHAR(255),
-    societe_gestion_id INTEGER,
+    societe_gestion_id BIGINT,
     CONSTRAINT fk_scpis_societe_gestion
         FOREIGN KEY (societe_gestion_id)
         REFERENCES societes_gestion(id)
 );
 
 CREATE TABLE IF NOT EXISTS taux_distribution (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     scpi_id BIGINT NOT NULL,
     annee INT NOT NULL,
     taux_distribution DECIMAL(5,2) NOT NULL,
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS taux_distribution (
 );
 
 CREATE TABLE IF NOT EXISTS decote_demembrement (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     scpi_id BIGINT NOT NULL,
     duree_annee INT NOT NULL,
     pourcentage DECIMAL(5,2) NOT NULL,
@@ -46,7 +46,7 @@ CREATE TABLE IF NOT EXISTS decote_demembrement (
 );
 
 CREATE TABLE IF NOT EXISTS valeurs_scpi (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     scpi_id BIGINT NOT NULL,
     annee INT NOT NULL,
     prix_part DECIMAL(10,2) NOT NULL,
@@ -57,34 +57,34 @@ CREATE TABLE IF NOT EXISTS valeurs_scpi (
 );
 
 CREATE TABLE IF NOT EXISTS localisations (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     pays VARCHAR(100),
     pourcentage NUMERIC(5,2),
-    scpi_id INTEGER,
+    scpi_id BIGINT,
     CONSTRAINT fk_localisation_scpi
         FOREIGN KEY (scpi_id)
         REFERENCES scpis(id)
 );
 
 CREATE TABLE IF NOT EXISTS secteurs (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     secteur VARCHAR(100),
     pourcentage NUMERIC(5,2),
-    scpi_id INTEGER,
+    scpi_id BIGINT,
     CONSTRAINT fk_secteurs_scpi
         FOREIGN KEY (scpi_id)
         REFERENCES scpis(id)
 );
 
 CREATE TABLE IF NOT EXISTS investisseurs (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     id_utilisateur VARCHAR(100),
     adresse VARCHAR(255)
 );
 
 CREATE TABLE IF NOT EXISTS patrimoines_immobiliers (
-    investisseur_id INTEGER NOT NULL,
-    scpi_id INTEGER NOT NULL,
+    investisseur_id BIGINT NOT NULL,
+    scpi_id BIGINT NOT NULL,
     PRIMARY KEY (investisseur_id, scpi_id),
     CONSTRAINT fk_patrimoine_investisseur
         FOREIGN KEY (investisseur_id)
@@ -95,8 +95,8 @@ CREATE TABLE IF NOT EXISTS patrimoines_immobiliers (
 );
 
 CREATE TABLE IF NOT EXISTS documents (
-    id SERIAL PRIMARY KEY,
-    investisseur_id INT,
+    id BIGSERIAL PRIMARY KEY,
+    investisseur_id BIGINT,
     nom VARCHAR(255),
     type VARCHAR(100),
     taille VARCHAR(50),
